pyproject.toml
[project]
name = "network-diagnostics-toolkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dnspython>=2.7.0",
    "httpx>=0.28.1",
    "typer>=0.15.2",
]
